Downloader - added new downloader to use serverlessworkflow.io instead of github - added utils for common tools functions - updated to latest schema version
