+++ src/lib/schema/common.json
@@ -1,5 +1,5 @@
 {
-  "$id": "https://serverlessworkflow.org/core/0.6/common.json",
+  "$id": "https://serverlessworkflow.io/schemas/0.6/common.json",
   "$schema": "http://json-schema.org/draft-07/schema#",
   "description": "Serverless Workflow specification - common schema",
   "type": "object",
