+++ src/lib/schema/events.json
@@ -1,5 +1,5 @@
 {
-  "$id": "https://serverlessworkflow.org/core/0.6/events.json",
+  "$id": "https://serverlessworkflow.io/schemas/0.6/events.json",
   "$schema": "http://json-schema.org/draft-07/schema#",
   "description": "Serverless Workflow specification - events schema",
   "type": "object",
