+++ src/lib/schema/functions.json
@@ -1,5 +1,5 @@
 {
-  "$id": "https://serverlessworkflow.org/core/0.6/functions.json",
+  "$id": "https://serverlessworkflow.io/schemas/0.6/functions.json",
   "$schema": "http://json-schema.org/draft-07/schema#",
   "description": "Serverless Workflow specification - functions schema",
   "type": "object",
