+++ src/lib/schema/retries.json
@@ -1,5 +1,5 @@
 {
-  "$id": "https://serverlessworkflow.org/core/0.6/retries.json",
+  "$id": "https://serverlessworkflow.io/schemas/0.6/retries.json",
   "$schema": "http://json-schema.org/draft-07/schema#",
   "description": "Serverless Workflow specification - retries schema",
   "type": "object",
